For TPUs/requirements.txt
pandas
numpy
transformers
wandb
torchinfo
torchmetrics
tqdm
matplotlib
scikit-learn
opencv-python
opencv-contrib-python
timm
# accimage               -- conda